requirements.txt
addict==2.4.0
beautifulsoup4==4.11.2
Brotli==1.0.9
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
chardet==5.1.0
charset-normalizer==3.0.1
chumpy==0.70
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
commandlines==0.4.1
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
dask==2023.3.1
decorator==4.4.2
ffmpeg-python==0.2.0
filelock==3.9.0
filetype==1.2.0
fonttools==4.38.0
fsspec==2023.3.0
future==0.18.3
gdown==4.6.4
hsh==1.1.0
idna==3.4
imageio==2.26.0
imageio-ffmpeg==0.4.8
joblib==1.2.0
json-tricks==3.16.1
kiwisolver==1.4.4
lazy_loader==0.1
llvmlite==0.39.1
locket==1.0.0
Markdown==3.4.1
markdown-it-py==2.2.0
matplotlib==3.7.0
mdurl==0.1.2
mmcv-full==1.7.1
mmdet==2.28.2
mmengine==0.6.0
-e git+https://github.com/open-mmlab/mmpose.git@21aeeb455fe198083cd3a16de9c9fc00bade8706#egg=mmpose
model-index==0.1.11
more-itertools==9.1.0
moviepy==1.0.3
munkres==1.1.4
mutagen==1.46.0
networkx==3.0
numba==0.56.4
numpy==1.23.5
opencv-python==4.7.0.72
openmim==0.3.6
ordered-set==4.1.0
packaging==23.0
pandas==1.5.3
partd==1.3.0
Pillow==9.4.0
proglog==0.1.10
protobuf==3.20.3
pycocotools==2.0.6
pycryptodomex==3.17
Pygments==2.14.0
PyMatting==1.1.8
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rich==13.3.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soupsieve==2.4
tabulate==0.9.0
tensorboardX==2.6
termcolor==2.2.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2023.2.28
toolz==0.12.0
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.6
waitress==2.1.2
websockets==10.4
wincertstore==0.2
xtcocotools==1.13
yapf==0.32.0
youtube-dl==2021.12.17
yt-dlp==2023.3.4
